p2p/enr: all predefined keys in one file
